Exposure time, temperature and rewritten file save
Exposure time now can be changed to one of the possible values on the
list.
You can now connect the arduino data logger to get temperature readings.
Adjusted the way the script handels new files, now the script does not
need to wait for p.lastImageFile
The script is set to take images in a shorter interval, altough it is
not optimized and it takes between 20 - 60 seconds to save the new image
(Depending on how full the RAM is I think?).