Bash script and major fixes
Added a bash script that checks for new files and renames them in the
correct direcotry. it is initiated and killed using matlab-process.

Got rid of the "start" function, the class is now nonblocking, moved
what remained from "start" to "connect".

To check for new files and get times, use "imageTimer". It is not as
good, and it might be improved (or removed), as its current use is
debugging only.

Fixed a lot of bugs - class now seems to work right, not intensively
tested.

Re-added plot. It seemed like the camera refused to capture images
without it. But now disconnect closes it.

Don't forcibly clear the class without disconnecting properly!

TODO: QA and documantation